+++ stem/descriptor/export.py
@@ -10,6 +10,11 @@ Toolkit for exporting descriptors to other formats.
 
   export_csv - Exports descriptors to a CSV
   export_csv_file - Writes exported CSV output to a file
+
+.. deprecated:: 1.7.0
+
+   This module will likely be removed in Stem 2.0 due to lack of usage. If you
+   use this modle please `let me know <https://www.atagar.com/contact/>`_.
 """
 
 import csv